ANN/examples/spin-up-with-variational-bayes/requirements.txt
absl-py==0.9.0
dm-haiku==0.0.1
jax==0.1.72
jaxlib==0.1.51
matplotlib==3.2.2
tensorflow==2.3.0
tensorflow_datasets==3.1.0
